hw-5

+++ WebSiteSolution/Application/Services/ApartmentService.cs
@@ -1,5 +1,7 @@
-﻿using Application.DTO;
+﻿using Application.Exceptions;
 using Application.Interfaces;
+using Application.Requests.ApartmentRequests;
+using Application.Responses;
 using AutoMapper;
 using Domain.Entities;
 using Infrastructure.Interfaces;
@@ -9,67 +11,61 @@ namespace Application.Services
     public class ApartmentService : IApartmentService
     {
         private readonly IApartmentRepository _apartmentRepository;
-        private readonly IUserRepository _userRepository;
+        private readonly IDealRepository _dealRepository;
+        private readonly IReviewRepository _reviewRepository;
         private readonly IMapper _mapper;
 
-        public ApartmentService(IApartmentRepository apartRepository, IUserRepository userRepository, IMapper mapper)
+        public ApartmentService(IApartmentRepository apartRepository, IReviewRepository reviewRepository, 
+            IDealRepository dealRepository, IMapper mapper)
         {
             _apartmentRepository = apartRepository;
-            _userRepository = userRepository;
+            _reviewRepository = reviewRepository;
+            _dealRepository = dealRepository;
             _mapper = mapper;
         }
 
-        public async Task<int> Add(ApartmentDTO apartment)
+        public async Task<int> Add(CreateApartmentRequest apartment)
         {
             var mappedApartment = _mapper.Map<Apartment>(apartment);
-            if (mappedApartment == null)
-            {
-                return -1;
-            }
-
-            var user = _userRepository.GetById(apartment.UserId);
-            if (user == null)
-            {
-                return -1;
-            }
-
             return await _apartmentRepository.Create(mappedApartment);
         }
 
         public async Task<bool> Delete(int id)
         {
+            var apartment = await _apartmentRepository.GetById(id);
+            if (apartment == null)
+                throw new NotFoundApplicationException($"Apartment with id {id} not found!");
+
+            await _reviewRepository.DeleteByApartmentId(id);
+            await _dealRepository.DeleteByApartmentId(id);
+
             return await _apartmentRepository.Delete(id);
         }
 
-        public async Task<IEnumerable<ApartmentDTO>> GetAll()
+        public async Task<IEnumerable<ApartmentResponse>> GetAll()
         {
             var aparts = await _apartmentRepository.GetAll();
-            var mappedAparts = aparts.Select(a => _mapper.Map<ApartmentDTO>(a)).ToList();
+            var mappedAparts = aparts.Select(a => _mapper.Map<ApartmentResponse>(a)).ToList();
             return mappedAparts;
         }
 
-        public async Task<ApartmentDTO> GetById(int id)
+        public async Task<ApartmentResponse> GetById(int id)
         {
             var apart = await _apartmentRepository.GetById(id);
-            var mappedAparts = _mapper.Map<ApartmentDTO>(apart);
-            return mappedAparts;
+            if (apart == null)
+            {
+                throw new NotFoundApplicationException($"Apartment with id {id} not found!");
+            }
+            
+            return _mapper.Map<ApartmentResponse>(apart);
         }
 
-        public async Task<bool> Update(ApartmentDTO apartment)
+        public async Task<bool> Update(UpdateApartmentRequest apartment)
         {
             if (apartment == null)
-            {
-                return false;
-            }
+                throw new NotFoundApplicationException($"Apartment not found!");
 
             var mappedApartment = _mapper.Map<Apartment>(apartment);
-
-            var user = _userRepository.GetById(apartment.UserId);
-            if (user == null)
-            {
-                return false;
-            }
-
             return await _apartmentRepository.Update(mappedApartment);
         }
     }
