+++ WebSiteSolution/Application/Services/ReviewService.cs
@@ -1,5 +1,7 @@
-﻿using Application.DTO;
+﻿using Application.Exceptions;
 using Application.Interfaces;
+using Application.Requests.ReviewRequests;
+using Application.Responses;
 using AutoMapper;
 using Domain.Entities;
 using Infrastructure.Interfaces;
@@ -17,43 +19,43 @@ namespace Application.Services
             _mapper = mapper;
         }
 
-        public async Task<int> Add(ReviewDTO review)
+        public async Task<int> Add(CreateReviewRequest review)
         {
             var mappedReview = _mapper.Map<Review>(review);
-            if (mappedReview != null)
-            {
-                await _reviewRepository.Create(mappedReview);
-                return mappedReview.Id;
-            }
-
-            return -1;
+            return await _reviewRepository.Create(mappedReview);
         }
 
         public async Task<bool> Delete(int id)
         {
+            var review = await _reviewRepository.GetById(id);
+            if (review == null)
+                throw new NotFoundApplicationException($"Review with id {id} not found!");
+
             return await _reviewRepository.Delete(id);
         }
 
-        public async Task<IEnumerable<ReviewDTO>> GetAll()
+        public async Task<IEnumerable<ReviewResponse>> GetAll()
         {
             var reviews = await _reviewRepository.GetAll();
-            var mappedReviews = reviews.Select(r => _mapper.Map<ReviewDTO>(r)).ToList();
+            var mappedReviews = reviews.Select(r => _mapper.Map<ReviewResponse>(r)).ToList();
             return mappedReviews;
         }
 
-        public async Task<ReviewDTO> GetById(int id)
+        public async Task<ReviewResponse> GetById(int id)
         {
             var review = await _reviewRepository.GetById(id);
-            var mappedReview = _mapper.Map<ReviewDTO>(review);
-            return mappedReview;
-        }
-
-        public async Task<bool> Update(ReviewDTO review)
-        {
             if (review == null)
             {
-                return false;
+                throw new NotFoundApplicationException($"Review with id {id} not found!");
             }
+            
+            return _mapper.Map<ReviewResponse>(review);
+        }
+
+        public async Task<bool> Update(UpdateReviewRequest review)
+        {
+            if (review == null)
+                throw new NotFoundApplicationException($"Review not found!");
 
             var mappedReview = _mapper.Map<Review>(review);
             return await _reviewRepository.Update(mappedReview);
