+++ WebSiteSolution/Application/Services/UserService.cs
@@ -1,5 +1,7 @@
-﻿using Application.DTO;
+﻿using Application.Exceptions;
 using Application.Interfaces;
+using Application.Requests.UserRequests;
+using Application.Responses;
 using AutoMapper;
 using Domain.Entities;
 using Infrastructure.Interfaces;
@@ -9,50 +11,62 @@ namespace Application.Services
     public class UserService : IUserService
     {
         private readonly IUserRepository _userRepository;
+        private readonly IApartmentRepository _apartmentRepository;
+        private readonly IDealRepository _dealRepository;
+        private readonly IReviewRepository _reviewRepository;
         private readonly IMapper _mapper;
 
-        public UserService(IUserRepository userRepository, IMapper mapper)
+        public UserService(IUserRepository userRepository, IApartmentRepository apartmentRepository, IDealRepository dealRepository,
+            IReviewRepository reviewRepository, IMapper mapper)
         {
             _userRepository = userRepository;
+            _apartmentRepository = apartmentRepository;
+            _dealRepository = dealRepository;
+            _reviewRepository = reviewRepository;
             _mapper = mapper;
         }
 
-        public async Task<int> Add(UserDTO user)
+        public async Task<int> Add(CreateUserRequest user)
         {
             var mappedUser = _mapper.Map<User>(user);
-            if (mappedUser == null)
-            {
-                return -1;
-            }
-
             return await _userRepository.Create(mappedUser);
         }
 
         public async Task<bool> Delete(int id)
         {
+            var user = await _userRepository.GetById(id);
+            if (user == null)
+                throw new NotFoundApplicationException($"User with id {id} not found!");
+
+            await _apartmentRepository.DeleteByOwnerId(id);
+            await _reviewRepository.DeleteByUserId(id);
+            await _dealRepository.DeleteByUserId(id);
+
             return await _userRepository.Delete(id);
         }
 
-        public async Task<IEnumerable<UserDTO>> GetAll()
+        public async Task<IEnumerable<UserResponse>> GetAll()
         {
             var users = await _userRepository.GetAll();
-            var mappedUsers = users.Select(u => _mapper.Map<UserDTO>(u)).ToList();
+            var mappedUsers = users.Select(u => _mapper.Map<UserResponse>(u)).ToList();
             return mappedUsers;
         }
 
-        public async Task<UserDTO?> GetById(int id)
+        public async Task<UserResponse> GetById(int id)
         {
             var user = await _userRepository.GetById(id);
-            var mappedUser = _mapper.Map<UserDTO>(user);
-            return mappedUser;
-        }
-
-        public async Task<bool> Update(UserDTO user)
-        {
             if (user == null)
             {
-                return false;
+                throw new NotFoundApplicationException($"User with id {id} not found!");
             }
+
+            return _mapper.Map<UserResponse>(user);
+        }
+
+        public async Task<bool> Update(UpdateUserRequest user)
+        {
+            if (user == null)
+                throw new NotFoundApplicationException($"User not found!");
 
             var mappedUser = _mapper.Map<User>(user);
             return await _userRepository.Update(mappedUser);
